+++ tests/integ/test_function.py
@@ -154,6 +154,7 @@ from snowflake.snowpark.types import (
 from tests.utils import TestData, Utils
 
 
+@pytest.mark.localtest
 def test_order(session):
     null_data1 = TestData.null_data1(session)
     assert null_data1.sort(asc(null_data1["A"])).collect() == [
@@ -200,6 +201,11 @@ def test_order(session):
     ]
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_current_date_and_time(session):
     df1 = session.sql("select current_date(), current_time(), current_timestamp()")
     df2 = session.create_dataframe([1]).select(
@@ -208,6 +214,11 @@ def test_current_date_and_time(session):
     assert len(df1.union(df2).collect()) == 1
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 @pytest.mark.parametrize("col_a", ["a", col("a")])
 def test_regexp_replace(session, col_a):
     df = session.create_dataframe(
@@ -229,6 +240,11 @@ def test_regexp_replace(session, col_a):
     assert res[0][0] == "lastname, firstname middlename"
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_regexp_extract(session):
     df = session.createDataFrame([["id_20_30", 10], ["id_40_50", 30]], ["id", "age"])
     res = df.select(regexp_extract("id", r"(\d+)", 1).alias("RES")).collect()
@@ -237,6 +253,11 @@ def test_regexp_extract(session):
     assert res[0]["RES"] == "30" and res[1]["RES"] == "50"
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 @pytest.mark.parametrize(
     "col_a, col_b, col_c", [("a", "b", "c"), (col("a"), col("b"), col("c"))]
 )
@@ -246,6 +267,11 @@ def test_concat(session, col_a, col_b, col_c):
     assert res[0][0] == "123"
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 @pytest.mark.parametrize(
     "col_a, col_b, col_c", [("a", "b", "c"), (col("a"), col("b"), col("c"))]
 )
@@ -255,6 +281,7 @@ def test_concat_ws(session, col_a, col_b, col_c):
     assert res[0][0] == "1,2,3"
 
 
+@pytest.mark.localtest
 @pytest.mark.parametrize("col_a", ["a", col("a")])
 def test_to_char(session, col_a):
     df = session.create_dataframe([[1]], schema=["a"])
@@ -262,12 +289,18 @@ def test_to_char(session, col_a):
     assert res[0][0] == "1"
 
 
+@pytest.mark.localtest
 def test_date_to_char(session):
     df = session.create_dataframe([[datetime.date(2021, 12, 21)]], schema=["a"])
     res = df.select(to_char(col("a"), "mm-dd-yyyy")).collect()
     assert res[0][0] == "12-21-2021"
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_format_number(session):
     # Create a dataframe with a column of numbers
     data = [
@@ -284,6 +317,11 @@ def test_format_number(session):
     assert res[2].VALUE_FORMATTED == "1.41"
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 @pytest.mark.parametrize("col_a, col_b", [("a", "b"), (col("a"), col("b"))])
 def test_months_between(session, col_a, col_b):
     df = session.create_dataframe(
@@ -293,6 +331,7 @@ def test_months_between(session, col_a, col_b):
     assert res[0][0] == 1.0
 
 
+@pytest.mark.localtest
 @pytest.mark.parametrize("col_a", ["a", col("a")])
 def test_cast(session, col_a):
     df = session.create_dataframe([["2018-01-01"]], schema=["a"])
@@ -301,6 +340,7 @@ def test_cast(session, col_a):
     assert cast_res[0][0] == try_cast_res[0][0] == datetime.date(2018, 1, 1)
 
 
+@pytest.mark.localtest
 @pytest.mark.parametrize("number_word", ["decimal", "number", "numeric"])
 def test_cast_decimal(session, number_word):
     df = session.create_dataframe([[5.2354]], schema=["a"])
@@ -309,18 +349,21 @@ def test_cast_decimal(session, number_word):
     )
 
 
+@pytest.mark.localtest
 def test_cast_map_type(session):
     df = session.create_dataframe([['{"key": "1"}']], schema=["a"])
     result = df.select(cast(parse_json(df["a"]), "object")).collect()
     assert json.loads(result[0][0]) == {"key": "1"}
 
 
+@pytest.mark.localtest
 def test_cast_array_type(session):
     df = session.create_dataframe([["[1,2,3]"]], schema=["a"])
     result = df.select(cast(parse_json(df["a"]), "array")).collect()
     assert json.loads(result[0][0]) == [1, 2, 3]
 
 
+@pytest.mark.localtest
 def test_startswith(session):
     Utils.check_answer(
         TestData.string4(session).select(col("a").startswith(lit("a"))),
@@ -329,6 +372,11 @@ def test_startswith(session):
     )
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_struct(session):
     df = session.createDataFrame([("Bob", 80), ("Alice", None)], ["name", "age"])
     # case sensitive
@@ -360,6 +408,11 @@ def test_struct(session):
     assert re.sub(r"\s", "", res[1].STRUCT) == '{"A":null,"B":"Alice"}'
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_strtok_to_array(session):
     # Create a dataframe
     data = [("a.b.c")]
@@ -368,6 +421,11 @@ def test_strtok_to_array(session):
     assert res[0] == "a" and res[1] == "b" and res[2] == "c"
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 @pytest.mark.parametrize(
     "col_a, col_b, col_c", [("a", "b", "c"), (col("a"), col("b"), col("c"))]
 )
@@ -377,6 +435,11 @@ def test_greatest(session, col_a, col_b, col_c):
     assert res[0][0] == 3
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 @pytest.mark.parametrize(
     "col_a, col_b, col_c", [("a", "b", "c"), (col("a"), col("b"), col("c"))]
 )
@@ -386,6 +449,11 @@ def test_least(session, col_a, col_b, col_c):
     assert res[0][0] == 1
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 @pytest.mark.parametrize("col_a, col_b", [("a", "b"), (col("a"), col("b"))])
 def test_hash(session, col_a, col_b):
     df = session.create_dataframe([[10, "10"]], schema=["a", "b"])
@@ -396,6 +464,11 @@ def test_hash(session, col_a, col_b):
     assert res[0][1] == 3622494980440108984
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_basic_numerical_operations_negative(session):
     # sqrt
     df = session.sql("select 4").to_df("a")
@@ -446,6 +519,11 @@ def test_basic_numerical_operations_negative(session):
     assert "'CEIL' expected Column or str, got: <class 'list'>" in str(ex_info)
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_basic_string_operations(session):
     # Substring
     df = session.sql("select 'a not that long string'").to_df("a")
@@ -531,6 +609,11 @@ def test_basic_string_operations(session):
     assert "'REVERSE' expected Column or str, got: <class 'list'>" in str(ex_info)
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_substring_index(session):
     """test calling substring_index with delimiter as string"""
     df = session.create_dataframe([[0, "a.b.c.d"], [1, ""], [2, None]], ["id", "s"])
@@ -551,6 +634,11 @@ def test_substring_index(session):
     assert respos[2][0] is None
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_substring_index_col(session):
     """test calling substring_index with delimiter as column"""
     df = session.create_dataframe([["a,b,c,d", ","]], ["s", "delimiter"])
@@ -562,6 +650,11 @@ def test_substring_index_col(session):
     assert reslit[0][0] == "b,c,d"
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_bitshiftright(session):
     # Create a dataframe
     data = [(65504), (1), (4)]
@@ -570,6 +663,11 @@ def test_bitshiftright(session):
     assert res[0][0] == 32752 and res[1][0] == 0 and res[2][0] == 2
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_bround(session):
     # Create a dataframe
     data = [(decimal.Decimal(1.235)), decimal.Decimal(3.5)]
@@ -580,6 +678,7 @@ def test_bround(session):
     assert str(res[0][0]) == "1" and str(res[1][0]) == "4"
 
 
+# Enable for local testing after addressing SNOW-850268
 def test_count_distinct(session):
     df = session.create_dataframe(
         [["a", 1, 1], ["b", 2, 2], ["c", 1, None], ["d", 5, None]]
@@ -603,6 +702,9 @@ def test_count_distinct(session):
     assert df.select(count_distinct(df["*"])).collect() == [Row(2)]
 
 
+@pytest.mark.skipif(
+    condition="config.getvalue('local_testing_mode')", reason="Testing builtin"
+)
 def test_builtin_avg_from_range(session):
     """Tests the builtin functionality, using avg()."""
     avg = builtin("avg")
@@ -643,6 +745,9 @@ def test_builtin_avg_from_range(session):
     assert res == expected
 
 
+@pytest.mark.skipif(
+    condition="config.getvalue('local_testing_mode')", reason="Testing builtin"
+)
 def test_call_builtin_avg_from_range(session):
     """Tests the builtin functionality, using avg()."""
     df = session.range(1, 10, 2).select(call_builtin("avg", col("id")))
@@ -685,6 +790,11 @@ def test_call_builtin_avg_from_range(session):
     assert res == expected
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_is_negative(session):
     td = TestData.string1(session)
 
@@ -817,6 +927,7 @@ def test_is_negative(session):
     assert "Invalid argument types for function 'IS_TIMESTAMP_TZ'" in str(ex_info)
 
 
+@pytest.mark.localtest
 def test_parse_json(session):
     assert TestData.null_json1(session).select(parse_json(col("v"))).collect() == [
         Row('{\n  "a": null\n}'),
@@ -832,6 +943,11 @@ def test_parse_json(session):
     ]
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_as_negative(session):
     td = TestData.string1(session)
 
@@ -1010,6 +1126,7 @@ def test_as_negative(session):
     )
 
 
+@pytest.mark.localtest
 def test_to_date_to_array_to_variant_to_object(session):
     df = (
         session.create_dataframe([["2013-05-17", 1, 3.14, '{"a":1}']])
@@ -1035,6 +1152,7 @@ def test_to_date_to_array_to_variant_to_object(session):
     assert df1.schema.fields[3].datatype == MapType(StringType(), StringType())
 
 
+@pytest.mark.localtest
 def test_to_binary(session):
     res = (
         TestData.test_data1(session)
@@ -1151,6 +1269,7 @@ def test_array_sort(session):
     Utils.check_answer(res, [Row(SORTED_A="[\n  null,\n  20,\n  10,\n  0\n]")])
 
 
+@pytest.mark.localtest
 def test_coalesce(session):
     # Taken from FunctionSuite.scala
     Utils.check_answer(
@@ -1169,6 +1288,11 @@ def test_coalesce(session):
     assert "'COALESCE' expected Column or str, got: <class 'list'>" in str(ex_info)
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_uniform(session):
     df = session.sql("select 1").to_df("a")
 
@@ -1203,6 +1327,11 @@ def test_uniform(session):
     assert decimal_int == decimal_decimal
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_uniform_negative(session):
     df = session.sql("select 1").to_df("a")
     with pytest.raises(SnowparkSQLException) as ex_info:
@@ -1210,6 +1339,7 @@ def test_uniform_negative(session):
     assert "Numeric value 'z' is not recognized" in str(ex_info)
 
 
+@pytest.mark.localtest
 def test_negate_and_not_negative(session):
     with pytest.raises(TypeError) as ex_info:
         TestData.null_data2(session).select(negate(["A", "B", "C"]))
@@ -1220,6 +1350,11 @@ def test_negate_and_not_negative(session):
     assert "'NOT_' expected Column or str, got: <class 'list'>" in str(ex_info)
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_random_negative(session):
     df = session.sql("select 1")
     with pytest.raises(SnowparkSQLException) as ex_info:
@@ -1227,6 +1362,11 @@ def test_random_negative(session):
     assert "Numeric value 'abc' is not recognized" in str(ex_info)
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_check_functions_negative(session):
     df = session.sql("select 1").to_df("a")
 
@@ -1255,6 +1395,11 @@ def test_parse_functions_negative(session):
     assert "'PARSE_XML' expected Column or str, got: <class 'list'>" in str(ex_info)
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_json_functions_negative(session):
     df = session.sql("select 1").to_df("a")
 
@@ -1274,6 +1419,11 @@ def test_json_functions_negative(session):
     )
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_to_filetype_negative(session):
     df = session.sql("select 1").to_df("a")
     # to_json
@@ -1287,6 +1437,11 @@ def test_to_filetype_negative(session):
     assert "'TO_XML' expected Column or str, got: <class 'list'>" in str(ex_info)
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_array_distinct(session):
     df = session.sql("select 1 A")
     df = df.withColumn(
@@ -1299,6 +1454,11 @@ def test_array_distinct(session):
     assert array[0] == 1 and array[1] == 2 and array[2] == 3
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_array_negative(session):
     df = session.sql("select 1").to_df("a")
 
@@ -1390,6 +1550,11 @@ def test_array_negative(session):
     )
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_object_negative(session):
     df = session.sql("select 1").to_df("a")
 
@@ -1423,6 +1588,11 @@ def test_object_negative(session):
     assert "'OBJECT_PICK' expected Column or str, got: <class 'list'>" in str(ex_info)
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_date_operations_negative(session):
     df = session.sql("select 1").to_df("a")
 
@@ -1435,6 +1605,7 @@ def test_date_operations_negative(session):
     assert "'DATEADD' expected Column or str, got: <class 'list'>" in str(ex_info)
 
 
+# TODO: enable for local testing after addressing SNOW-850263
 def test_date_add_date_sub(session):
     df = session.createDataFrame(
         [("2019-01-23"), ("2019-06-24"), ("2019-09-20")], ["date"]
@@ -1450,12 +1621,22 @@ def test_date_add_date_sub(session):
     assert res[2].DATE == datetime.date(2019, 9, 16)
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_daydiff(session):
     df = session.createDataFrame([("2015-04-08", "2015-05-10")], ["d1", "d2"])
     res = df.select(daydiff(to_date(df.d2), to_date(df.d1)).alias("diff")).collect()
     assert res[0].DIFF == 32
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_get_negative(session):
     df = session.sql("select 1").to_df("a")
 
@@ -1464,6 +1645,11 @@ def test_get_negative(session):
     assert "'GET' expected Column, int or str, got: <class 'list'>" in str(ex_info)
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_array_generate_range(session):
     df = session.createDataFrame([(-2, 2)], ["C1", "C2"])
     Utils.check_answer(
@@ -1494,6 +1680,11 @@ def test_array_generate_range(session):
     )
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_sequence_negative(session):
     df = session.sql("select 1").to_df("a")
 
@@ -1502,6 +1693,11 @@ def test_sequence_negative(session):
     assert "'SEQUENCE' expected Column or str, got: <class 'list'>" in str(ex_info)
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_sequence(session):
     df = session.createDataFrame([(-2, 2)], ["C1", "C2"])
     Utils.check_answer(
@@ -1553,6 +1749,11 @@ def test_sequence(session):
     )
 
 
+@pytest.mark.xfail(
+    condition="config.getvalue('local_testing_mode')",
+    raises=NotImplementedError,
+    strict=True,
+)
 def test_array_unique_agg(session):
     def _result_str2lst(result):
         col_str = result[0][0]